requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
PyMySQL==1.1.0
cryptography==41.0.7
SQLAlchemy==2.0.23
email-validator==2.1.0
python-dotenv==1.0.0
hypothesis==6.92.1
pytest==7.4.3
pytest-flask==1.3.0
